requirements.txt
discord.py>=2.3,<3
python-dotenv>=1.0,<2
fastapi>=0.110,<1
uvicorn[standard]>=0.29,<1
jinja2>=3.1,<4
python-multipart>=0.0.9,<1
markdown>=3.5,<4
itsdangerous>=2.1,<3
python-frontmatter>=1.1,<2
boto3>=1.34,<2

# Dev / CI
ruff>=0.15,<1
pytest>=8,<9
pytest-cov>=7,<8
pytest-xdist>=3.5,<4
pytest-asyncio>=0.23,<1
httpx>=0.27,<1
moto>=5,<6
